split event-service into multiple smaller services
