+++ src/event/event.module.ts
@@ -1,6 +1,5 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { EventController } from './event.controller';
-import { EventService } from './event.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { EventEntity } from './infrastructure/persistence/relational/entities/event.entity';
 import { TenantModule } from '../tenant/tenant.module';
@@ -17,6 +16,11 @@ import { UserModule } from '../user/user.module';
 import { GroupModule } from '../group/group.module';
 import { EventMailModule } from '../event-mail/event-mail.module';
 import { BlueskyModule } from '../bluesky/bluesky.module';
+import { EventManagementService } from './services/event-management.service';
+import { EventQueryService } from './services/event-query.service';
+import { EventRecommendationService } from './services/event-recommendation.service';
+import { EventDiscussionService } from './services/event-discussion.service';
+
 @Module({
   imports: [
     TypeOrmModule.forFeature([EventEntity]),
@@ -33,12 +37,20 @@ import { BlueskyModule } from '../bluesky/bluesky.module';
   ],
   controllers: [EventController],
   providers: [
-    EventService,
+    EventManagementService,
+    EventQueryService,
+    EventRecommendationService,
+    EventDiscussionService,
     FilesS3PresignedService,
     EventListener,
     ZulipService,
     EventRoleService,
   ],
-  exports: [EventService],
+  exports: [
+    EventManagementService,
+    EventQueryService,
+    EventRecommendationService,
+    EventDiscussionService,
+  ],
 })
 export class EventModule {}
