+++ src/test/mocks/event-recommendation-mocks.ts
@@ -1,0 +1,8 @@
+import { mockEvents } from './event-mocks';
+
+export const mockEventRecommendationService = {
+  showRandomEvents: jest.fn().mockResolvedValue(mockEvents),
+  findRecommendedEventsForGroup: jest.fn().mockResolvedValue(mockEvents),
+  findRandomEventsForGroup: jest.fn().mockResolvedValue(mockEvents),
+  findRandom: jest.fn().mockResolvedValue(mockEvents),
+};
