+++ src/test/mocks/index.ts
@@ -1,5 +1,4 @@
 export * from './zulip-mocks';
 export * from './mocks';
-export * from './event-mocks';
 export * from './group-mocks';
 export * from './user-mocks';
